detector/requirements.txt
Pillow
tflite_runtime
redis